requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
elasticsearch
p115client
